apply modifications in regex logic

+++ controllers/catagory.controler.js
@@ -1,31 +1,25 @@
 const catagory = require("../models/catagory.model.js");
 const faq = require("../models/faq.model.js");
 const { validateAdmin } = require("../utils/jwt.js");
-const {emptyFieldFinder,successResponse,errorResponse} = require('../utils/constants.js')
+const {emptyFieldFinder,successResponse,errorResponse,validateRegex} = require('../utils/constants.js')
 const {Op} = require('sequelize')
 
 const createCatagory = async (req, res) => {
   const { name, description } = req.body;
+  const validateRole = await validateAdmin(req.headers.auth_token);
   if (!(name && description)) {
     const emptyField = emptyFieldFinder({name,description})
     return errorResponse(res,400,`${emptyField} should not be empty`)
   }
-
+  if (!validateRole) {
+    return errorResponse(res,401,"Only admin can create a catagory")
+  }
   try {
-    const validateRole = await validateAdmin(req.headers.auth_token);
-    if (validateRole) {
-      try {
-        await catagory.create({
-          name,
-          description,
-        });
-        successResponse(res,200,'catagory has been created successfully')
-      } catch (error) {
-        errorResponse(res,500,error.message);
-      }
-    } else {
-      errorResponse(res,200,'Only admin can create catagory');
-    }
+    await catagory.create({
+      name,
+      description,
+    });
+    successResponse(res,200,'catagory has been created successfully')
   } catch (error) {
     errorResponse(res,500,error.message);
   }
@@ -35,6 +29,9 @@ const getSingleCatagory = async(req,res) => {
   const {id} = req.query
   if (!id) {
    return errorResponse(res,400,"Please provide catagory ID")
+  }
+  if (!(validateRegex(id))) {
+    return errorResponse(res,400,"Please provide a valid catagory Id")
   }
   try {
     const resp = await catagory.findOne({
@@ -51,7 +48,6 @@ const getSingleCatagory = async(req,res) => {
     console.log(error)
     errorResponse(res,500,error.message)
   }
-
 }
 
 const obtainCatagories = async (req, res) => {
@@ -78,7 +74,6 @@ const obtainCatagories = async (req, res) => {
       }
     }
     const count = await catagory.count()
-    
     const catagories = await catagory.findAll(findOptions);
     successResponse(res,200,catagories,count)
   } catch (error) {
@@ -99,6 +94,9 @@ const updateCatagory = async (req, res) => {
   if (!validateRole) {
     return errorResponse(res,401,'Only admin can create catagory')
   }
+  if (!(validateRegex(catagory_id))) {
+    return errorResponse(res,400,"Please provide a valid catagory Id")
+  }
   try {
     const data = await catagory.findOne({where : {id : catagory_id}})
     if (data) {
@@ -111,35 +109,44 @@ const updateCatagory = async (req, res) => {
     console.log(error)
     errorResponse(res,500,error.message)
   }
+  
 };
 
 const deleteCatagory = async (req, res) => {
   const { catagory_id } = req.query;
+  const validateRole =  await validateAdmin(req.headers.auth_token)
   if (!catagory_id) {
     return  errorResponse(res,400,"please provide catagory id to proceed")
   } 
-  const validateRole =  await validateAdmin(req.headers.auth_token)
   if (!validateRole) {
     return errorResponse(res,401,"Only admin can delete a catagory")
   }
+  if (!(validateRegex(catagory_id))) {
+    return errorResponse(res,400,"Please provide a valid catagory Id")
+  }
   try {
     const resp = await catagory.destroy({ where: { id: catagory_id } });
-    resp? successResponse(res,200,"catagory has been deleted successfully") : errorResponse(res,404,"catagory does not exist")
+    resp? successResponse(res,200,"catagory has been deleted successfully") 
+    : errorResponse(res,404,"catagory does not exist")
   } catch (error) {
     console.log(error);
     errorResponse(res,500,error.message);
   }
+  
 };
 
 const bulkCatagoryDelete = async(req,res) => {
     const {ids} = req.body
-    // if (!(Array.isArray(ids) && ids.length > 0)) {
-    //   return !Array.isArray(ids)? errorResponse(res,400,"Ids should be array in bulk delete") : 
-    //          errorResponse(res,400,"Please select atleast one catagory to proceed")
-    // }
     const validateRole = await validateAdmin(req.headers.auth_token)
+     if (!(Array.isArray(ids) && ids.length > 0)) {
+       return !Array.isArray(ids)? errorResponse(res,400,"Ids should be array in bulk delete") : 
+       errorResponse(res,400,"Please select atleast one catagory to proceed")
+     }
     if (!validateRole) {
     return errorResponse(res,401,"Only admin can delete catagories")
+    }
+    if (!(validateRegex(ids))) {
+      return errorResponse(res,400,"Please provide a valid catagory IDs")  
     }
     try {
       const resp = await catagory.destroy({where : {id : ids}})
