+++ controllers/faq.controller.js
@@ -1,18 +1,21 @@
 const faq = require("../models/faq.model.js");
 const catagory = require("../models/catagory.model.js");
-const { Op, where } = require("sequelize");
+const { Op } = require("sequelize");
 const {validateAdmin} = require('../utils/jwt.js')
-const {emptyFieldFinder, errorResponse, successResponse} = require('../utils/constants.js');
+const {emptyFieldFinder, errorResponse, successResponse, validateRegex} = require('../utils/constants.js');
 
 const createFaq = async (req, res) => {
   const { question, answer, catagory_id, } = req.body;
+  const validateRole = await validateAdmin(req.headers.auth_token)
   if (!(question && answer && catagory_id)) {
     const emptyField = emptyFieldFinder({question,answer,catagory_id})
     return errorResponse(res,400,`${emptyField} should not be empty`)
   }
-  const validateRole = validateAdmin(req.headers.auth_token)
   if (!validateRole) {
     return errorResponse(res,401,"Only admin can create FAQs")
+  }
+  if (!(validateRegex(catagory_id))) {
+    return errorResponse(res,400,"Please provide a valid catagory Id")
   }
     try {
       const resp = await catagory.findOne({where : {id : catagory_id}})
@@ -26,7 +29,6 @@ const createFaq = async (req, res) => {
       }else{
         errorResponse(res,404,"Catagory does not exist")
       }
-        
       } catch (error) {
         console.log(error);
         errorResponse(res,500,error.message)
@@ -37,6 +39,9 @@ const getSingleFaq = async(req,res) => {
   const {id} = req.query
   if (!id) {
     return errorResponse(res,400,"Please provide FAQ Id to proceed")
+  }
+  if (!(validateRegex(id))) {
+    return errorResponse(res,400,"Please provide Id in valid format") 
   }
   try {
     const resp = await faq.findOne({
@@ -86,14 +91,13 @@ const getFaq = async (req, res) => {
 
 const updateFaq = async (req, res) => {
   const { catagory_id, question, answer, id } = req.body;
-  const regex = /^\d+$/
   if (!(catagory_id && question && answer && id)) {
     const emptyField = emptyFieldFinder({catagory_id,question,answer,id})
     return errorResponse(res,400,`${emptyField} should not be empty`)
   }
-  if (!(regex.test(id) && regex.test(catagory_id))) {
-    return !regex.test(id)? errorResponse(res,400,"Please provide Id in correct format") 
-    : errorResponse(res,400,"Please provide catagory Id in correct format")
+  if (!(validateRegex(catagory_id) && validateRegex(id))) {
+    return !validateRegex(id)? errorResponse(res,400,"Please provide a valid FAQ Id ") 
+    : errorResponse(res,400,"Please provide a valid catagory Id")
   }
   const validateRole = await validateAdmin(req.headers.auth_token)
   if (!validateRole) {
@@ -112,17 +116,19 @@ const updateFaq = async (req, res) => {
       console.log(error)
       errorResponse(res,500,error.message)
      } 
-    
 };
 
 const deleteFaq = async (req, res) => {
+    const validateRole =  await validateAdmin(req.headers.auth_token)
     const {id} = req.query
     if (!id) {
         return errorResponse(res,400,"please select FAQ Id to delete")
     }
-    const validateRole =  await validateAdmin(req.headers.auth_token)
     if (!validateRole) {
     return errorResponse(res,401,"Only admin can delete FAQs")
+    }
+    if (!(validateRegex(id))) {
+      return errorResponse(res,400,"Please provide a valid FAQ Id")
     }
     try {
         const resp = await faq.destroy({where : {id}})
@@ -135,10 +141,17 @@ const deleteFaq = async (req, res) => {
 };
 
 const bulkFaqDelete = async(req,res) => {
-    const {ids} = req.body
     const validateRole = await validateAdmin(req.headers.auth_token)
+    const {ids} = req.body
+    if (!(Array.isArray(ids) && ids.length > 0)) {
+      return !Array.isArray(ids)? errorResponse(res,400,"Ids should be array in bulk delete") : 
+      errorResponse(res,400,"Please select atleast one FAQ to proceed")
+    }
     if (!validateRole) {
     return errorResponse(res,401,"Only admin can delete FAQs")
+    }
+    if (!(validateRegex(ids))) {
+      return errorResponse(res,400,"Please provide valid FAQ IDs")
     }
     try {
         const resp = await faq.destroy({where : {id : ids}})
